# Implement static extraction (SSR) of Twind styles