--- a/assets/twind.js
+++ b/assets/twind.js
@@ -1,8 +1,0 @@
-import { setup } from 'https://cdn.skypack.dev/twind/shim'
-import typography from 'https://cdn.skypack.dev/@twind/typography'
-
-setup({
-  plugins: {
-    ...typography(),
-  },
-})