@mtsknn 2021-06-20
Performance grew from 99 points to 100 points when measured at [web.dev/measure](https://web.dev/measure/). 🥳 I ran the tool only once before and after merging the PR, but still:

| Metric                   | Before | After |
| ------------------------ | ------ | ----- |
| First Contentful Paint   | 1.3 s  | 0.8 s |
| Time to Interactive      | 2.0 s  | 0.8 s |
| Speed Index              | 2.3 s  | 0.9 s |
| Total Blocking Time      | 110 ms | 0 ms  |
| Largest Contentful Paint | 1.9 s  | 0.8 s |
| Cumulative Layout Shift  | 0      | 0     |

💪