@mtsknn 2021-06-28
[Related web.dev article: _Extract critical CSS_.](https://web.dev/extract-critical-css/)

A relevant quote:

> To minimize the number of roundtrips to first render, aim to keep above-the-fold content under 14 KB (compressed).

Currently all of our pages weigh less than 14 KB &ndash; not just above-the-fold content but a whole single page! &ndash; so inlining all CSS is okay from this viewpoint.

Another quote:

> Inlining also has some downsides in that it prevents the browser from caching the CSS for reuse on subsequent page loads, so it's best to use it sparingly.

Maybe the best combo would be:

- Inlined CSS (this PR) &rarr; the first page loads and renders fast.
- Using Ajax to load the contents of subsequent pages (i.e. instead of doing full page refreshes, the next page's contents are retrieved via Ajax and updated to the content area using JS) &rarr; it doesn't matter that the CSS is not cached.

But I just realized that the second point wouldn't work right now because every page has unique CSS...

Alternatively, maybe https://instant.page/ could be used for faster subsequent page loads.